requirements.txt
alembic==1.5.3
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==3.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
Pillow==8.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
toml==0.10.2
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
gunicorn
psycopg2